+++ src/main/java/gym_solution/demo/service/serviceImpl/PaymentServiceImpl.java
@@ -2,6 +2,7 @@ package gym_solution.demo.service.serviceImpl;
 
 import gym_solution.demo.dto.PaymentDTO;
 import gym_solution.demo.dto.response.PaymentResponseDTO;
+import gym_solution.demo.exception.NotFoundException;
 import gym_solution.demo.mapper.PaymentMapper;
 import gym_solution.demo.model.Payment;
 import gym_solution.demo.model.Subscription;
@@ -10,6 +11,7 @@ import gym_solution.demo.repository.SubscriptionRepository;
 import gym_solution.demo.service.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -24,10 +26,11 @@ public class PaymentServiceImpl implements PaymentService {
     private final PaymentMapper paymentMapper;
 
     @Override
+    @Transactional
     public PaymentResponseDTO addPayment(PaymentDTO paymentDTO) {
         Subscription subscription = this.subscriptionRepository.findById(paymentDTO.
                         getSubscriptionId()).
-                orElseThrow(() -> new RuntimeException("subscription is not found"));
+                orElseThrow(() -> new NotFoundException("Subscription with id " + paymentDTO.getSubscriptionId() + " not found"));
 
         Payment payment = this.paymentMapper.toPayment(paymentDTO);
         payment.setSubscription(subscription);
@@ -36,9 +39,10 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
+    @Transactional
     public PaymentResponseDTO updatePayment(PaymentResponseDTO paymentResponseDTO) {
 
-        Payment payment = this.paymentRepository.findById(paymentResponseDTO.getPaymentId()).orElseThrow(() -> new RuntimeException("payment is not found"));
+        Payment payment = this.paymentRepository.findById(paymentResponseDTO.getPaymentId()).orElseThrow(() -> new NotFoundException("Payment with id " + paymentResponseDTO.getPaymentId() + " not found"));
         payment.setAmount(paymentResponseDTO.getAmount());
         payment.setDate(paymentResponseDTO.getDate());
         payment.setMethod(paymentResponseDTO.getMethod());
@@ -47,13 +51,17 @@ public class PaymentServiceImpl implements PaymentService {
     }
 
     @Override
+    @Transactional
     public void deletePaymentById(Long id) {
+        if (!this.paymentRepository.existsById(id)) {
+            throw new NotFoundException("Payment with id " + id + " not found");
+        }
         this.paymentRepository.deleteById(id);
     }
 
     @Override
     public PaymentResponseDTO findPaymentById(Long id) {
-        return this.paymentMapper.toResponse(this.paymentRepository.findById(id).orElseThrow(() -> new RuntimeException("payment is not found")));
+        return this.paymentMapper.toResponse(this.paymentRepository.findById(id).orElseThrow(() -> new NotFoundException("Payment with id " + id + " not found")));
     }
 
     @Override
